modularizing the BVN scratch code
