+++ _rmd/extra_loss_moments/scratch_worked_ex1.py
@@ -1,0 +1,3 @@
+"""
+## (1.5) Worked examples
+"""
